repetition detection

+++ include/types.h
@@ -13,7 +13,7 @@ typedef struct {
 	uint64_t bitboard[12];
 	int board[64];
 	uint64_t hash;
-	int en_passant;
+	int8_t en_passant;
 	unsigned int castle : 4;
 	unsigned int side : 1;
 } Pos;
